requirements.txt
python-documentcloud==3.1.0
requests
